嘤嘤怪杀手可以@受害者了

+++ repeater.cpp
@@ -9,7 +9,7 @@ using namespace CQ;
 
 struct MessageDict
 {
-	int fromGroup = 0;
+	long long fromGroup = 0;
 	vector<string> messageDict;
 	vector<string> repeatedMessageDict;
 };
